[frontend] Move ProjectController#rebuild_time to it's own controller
* ProjectController#rebuild_time -> Projects::RebuildTimesController#show
* ProjectController#rebuild_time_png -> Projects::RebuildTimesController#rebuild_time_png
* Simplified before filters
* Moved _tabs partial to webui/shared/projects/_tabs

This is a first step to refactor the ProjectController to make it more SOLID.
