+++ src/api/app/helpers/webui/project_helper.rb
@@ -74,7 +74,7 @@ module Webui::ProjectHelper
   def rebuild_time_col(package)
     return '' if package.blank?
     btime = @timings[package][0]
-    link_to(h(package), controller: :package, action: :show, project: @project, package: package) + ' ' + format_seconds(btime)
+    link_to(h(package), controller: '/webui/package', action: :show, project: @project, package: package) + ' ' + format_seconds(btime)
   end
 
   def short_incident_name(incident)
